Completed version of openoffice_document_macro
