+++ data/exploits/openoffice_document_macro/Basic/Standard/Module1.xml
@@ -2,7 +2,5 @@
 <!DOCTYPE script:module PUBLIC "-//OpenOffice.org//DTD OfficeDocument 1.0//EN" "module.dtd">
 <script:module xmlns:script="http://openoffice.org/2000/script" script:name="Module1" script:language="StarBasic">REM  *****  BASIC  *****
 
-Sub OnLoad
-MsgBox &quot;Auto1111?&quot;
-End Sub
+CODEGOESHERE
 </script:module>
